Add rating history and review score histogram to stats page (#16)
Add rating history and review score histogram to the stats page, as well as the beginnings of an API to request all the information for the 'stats' page, instead of baking it into the html using templates. Also includes a general refactor to add 'services' instead of doing everything in the controllers, and a lot of clean up that was blocking me before.