Add puzzle history page, and save user's current 'next puzzle' unless they skip it (#26)
* Add Puzzle History page, which shows all the unique puzzles reviewed,
ordered by date. We should include new puzzles at some point, but we
should probably make the next puzzle the user sees fixed and force them
to skip it, and then register a review for it with the new difficulty of
'skipped', and then they'll automatically show up on this page.
* Save the 'next random puzzle' the user is currently looking at, and add
a skip button for if they really don't want to do it.