Add option to change the review order with three options (#32)
Add option to change the review order with three options: DueTime (the current mode), PuzzleRating (easier puzzles first), or Random (just show reviews in a random order). It might make sense to change the default mode to 'PuzzleRating' for the next version if it works well, since the current option of DueTime is a bit random.